imapclient: split into per-command files
