+++ imapclient/append.go
@@ -1,0 +1,43 @@
+package imapclient
+
+import (
+	"io"
+)
+
+// Append sends an APPEND command.
+//
+// The caller must call AppendCommand.Close.
+func (c *Client) Append(mailbox string, size int64) *AppendCommand {
+	// TODO: flag parenthesized list, date/time string
+	cmd := &AppendCommand{}
+	cmd.enc = c.beginCommand("APPEND", cmd)
+	cmd.enc.SP().Mailbox(mailbox).SP()
+	cmd.wc = cmd.enc.Literal(size)
+	return cmd
+}
+
+// AppendCommand is an APPEND command.
+//
+// Callers must write the message contents, then call Close.
+type AppendCommand struct {
+	cmd
+	enc *commandEncoder
+	wc  io.WriteCloser
+}
+
+func (cmd *AppendCommand) Write(b []byte) (int, error) {
+	return cmd.wc.Write(b)
+}
+
+func (cmd *AppendCommand) Close() error {
+	err := cmd.wc.Close()
+	if cmd.enc != nil {
+		cmd.enc.end()
+		cmd.enc = nil
+	}
+	return err
+}
+
+func (cmd *AppendCommand) Wait() error {
+	return cmd.cmd.Wait()
+}
