+++ imapclient/capability.go
@@ -1,0 +1,37 @@
+package imapclient
+
+import (
+	"fmt"
+
+	"github.com/emersion/go-imap/v2/internal/imapwire"
+)
+
+// Capability sends a CAPABILITY command.
+func (c *Client) Capability() *CapabilityCommand {
+	cmd := &CapabilityCommand{}
+	c.beginCommand("CAPABILITY", cmd).end()
+	return cmd
+}
+
+// CapabilityCommand is a CAPABILITY command.
+type CapabilityCommand struct {
+	cmd
+	caps map[string]struct{}
+}
+
+func (cmd *CapabilityCommand) Wait() (map[string]struct{}, error) {
+	err := cmd.cmd.Wait()
+	return cmd.caps, err
+}
+
+func readCapabilities(dec *imapwire.Decoder) (map[string]struct{}, error) {
+	caps := make(map[string]struct{})
+	for dec.SP() {
+		var name string
+		if !dec.ExpectAtom(&name) {
+			return caps, fmt.Errorf("in capability-data: %v", dec.Err())
+		}
+		caps[name] = struct{}{}
+	}
+	return caps, nil
+}
